+++ service_catalog/models/approval_workflow.py
@@ -1,8 +1,8 @@
-from django.db.models import CharField, ForeignKey, ManyToManyField, CASCADE
-
+from django.db.models import CharField, ForeignKey, ManyToManyField, CASCADE, BooleanField
+from hashlib import sha256
 from Squest.utils.squest_model import SquestModel
-from profiles.models import AbstractScope, GlobalScope
-from service_catalog.models import TowerSurveyField, ApprovalStep
+from profiles.models import AbstractScope, GlobalScope, Scope
+from service_catalog.models import TowerSurveyField, ApprovalStep, RequestState, Request
 
 
 class ApprovalWorkflow(SquestModel):
@@ -34,6 +34,14 @@ class ApprovalWorkflow(SquestModel):
         return all_fields - used_fields
 
     @property
+    def hash(self):
+        string = f"{self.enabled}_"
+        for step in self.approval_steps.all().order_by('id'):
+            string += f"{step.hash}_"
+        return int(sha256(string.encode("utf-8")).hexdigest(),16) % 2 ** 31
+
+
+    @property
     def first_step(self):
         first_step = ApprovalStep.objects.filter(approval_workflow=self, position=0)
         if first_step.exists():
@@ -46,19 +54,60 @@ class ApprovalWorkflow(SquestModel):
     def instantiate(self):
         from service_catalog.models import ApprovalStepState
         from service_catalog.models import ApprovalWorkflowState
-        new_approval_workflow_state = ApprovalWorkflowState.objects.create(approval_workflow=self)
+        new_approval_workflow_state = ApprovalWorkflowState.objects.create(approval_workflow=self, hash=self.hash)
         for approval_step in self.approval_steps.all():
-            new_app_workflow_state = ApprovalStepState.objects.create(
+            current_step_state = ApprovalStepState.objects.create(
                 approval_workflow_state=new_approval_workflow_state,
                 approval_step=approval_step)
             if approval_step.position == 0:
-                new_approval_workflow_state.current_step = new_app_workflow_state
+                new_approval_workflow_state.current_step = current_step_state
         new_approval_workflow_state.save()
         return new_approval_workflow_state
 
+    def _get_all_requests_that_should_use_workflow(self):
+        if not self.enabled:
+            return Scope.objects.none()
+        scopes_already_assigned_to_another_workflow = Scope.objects.filter(
+            approval_workflows__operation=self.operation,
+            approval_workflows__enabled=True
+        ).exclude(
+            approval_workflows__id=self.id
+        )
+        expanded_scopes = self.scopes.expand().exclude(id__in=scopes_already_assigned_to_another_workflow.values("id"))
+
+        if self.scopes.exists():
+            return Request.objects.filter(
+                operation=self.operation,
+                instance__quota_scope__id__in=expanded_scopes,
+                state=RequestState.SUBMITTED
+            )
+        else:
+            return Request.objects.filter(
+                operation=self.operation,
+                state=RequestState.SUBMITTED
+            ).exclude(
+                instance__quota_scope__id__in=scopes_already_assigned_to_another_workflow.expand()
+            )
+
+    def _get_request_using_workflow(self):
+        return Request.objects.filter(approval_workflow_state__approval_workflow=self, state=RequestState.SUBMITTED)
+
+    def _get_request_using_workflow_with_wrong_version(self):
+        return self._get_request_using_workflow().exclude(approval_workflow_state__hash=self.hash)
+
+    def _get_request_using_workflow_with_good_version(self):
+        return self._get_request_using_workflow().filter(approval_workflow_state__hash=self.hash)
+
+    def _get_request_to_reset(self):  # TODO Add tests
+        return self._get_all_requests_that_should_use_workflow() \
+            .exclude(id__in=self._get_request_using_workflow_with_good_version().values_list("id", flat=True)) \
+            | self._get_request_using_workflow_with_wrong_version() \
+            | self._get_request_using_workflow() \
+                .exclude(id__in=self._get_all_requests_that_should_use_workflow().values_list("id", flat=True))
+
     def reset_all_approval_workflow_state(self):
-        for approval_workflow_state in self.approval_workflow_states.all():
-            approval_workflow_state.reset()
+        for request in self._get_request_to_reset().distinct():
+            request.setup_approval_workflow()
 
     def get_scopes(self):
         scopes = AbstractScope.objects.none()